`Dry::CLI::Utils::Files::MemoryFileSystem::Path` => `Dry::CLI::Utils::Path`
